refactor: split spotify client into specialized modules
